Make ClientTest#deleteClient pass by expecting an Exception.
We do this since we want to keep the @ManyToOne mapping between
Invoice -> Client. But deleting a client won't work, since there's
a reference to a Client in the invoices table, and that
reference would be nulled (unpermitted, throws an exception).

This could however be seen as expected behavior, and can be
handled in the application logic later.